Move models library into cyclops
