+++ cyclops/models/neural_nets/lstm.py
