+++ cyclops/models/util.py
